Move some services into WorkerSharedGlobalServices
